Added some validation for phoneNumber

+++ BloodBankApp/Services/Interfaces/IAvailabilityService.cs
@@ -7,7 +7,7 @@ namespace BloodBankApp.Services.Interfaces
     {
         Task<bool> UsernameIsTaken(string username);
         Task<bool> PersonalNumberIsTaken(int personalNumber);
-        Task<bool> PersonalNumberIsTaken(Guid id, int personalNumber);
+        Task<bool> PersonalNumberIsTaken(Guid id, long personalNumber);
 
         Task<bool> HospitalCodeIsTaken(string hospitalCode);
         Task<bool> DonorCodeIsTaken(string codeValue);
